HW_2/requirements.txt
torch
torchvision
numpy
einops
pillow